requirements.txt
Flask==2.3.2
google-auth==2.22.0
google-api-python-client==2.82.0
gunicorn==20.1.0
python-dotenv==1.0.0